requirements.txt
tensorflow==2.3.1
streamlit==0.74.1
numpy==1.18.5
pillow==8.0.1